refactor: imporve the code structure of mp4-utils.
