+++ packages/av-cliper/src/mp4-utils/__tests__/mp4-utils.test.ts
@@ -1,7 +1,7 @@
-import './mock'
+import '../../__tests__/mock'
 import { beforeAll, describe, expect, test, vi } from 'vitest'
 import mp4box from '@webav/mp4box.js'
-import { file2stream } from '../mp4-utils'
+import { file2stream } from '..'
 
 beforeAll(() => {
   vi.useFakeTimers()
